pyproject.toml
[project]
name = "manta"
version = "0.1.0"
authors = [
  { name="Fischer Moseley", email="fischerm@mit.edu" },
]
description = "An In-Situ Debugging Tool for Programmable Hardware"
readme = "README.md"
dependencies = [
  "amaranth[builtin-yosys]@git+https://github.com/amaranth-lang/amaranth",
  "PyYAML",
  "pyserial",
  "pyvcd",
  "scapy"
]

requires-python = ">=3.7"

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "mkdocs-material",
  "amaranth_boards@git+https://github.com/amaranth-lang/amaranth-boards"
]

[project.urls]
"Homepage" = "https://github.com/fischermoseley/manta"

[project.scripts]
manta = "manta:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"